pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Agentic RAG workflow for AI startup research and investment reports"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.0,<0.4",
    "langchain-community>=0.3.0,<0.4",
    "langchain-openai>=0.2.0,<0.3",
    "langgraph>=0.2.0,<0.3",
    "langgraph-checkpoint-sqlite>=2.0.0",
    "faiss-cpu>=1.8.0,<2",
    "tavily-python>=0.5.0",
    "pydantic>=2.0.0,<3",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.7.0",
    "openai>=1.0.0",
    "pypdf>=5.0.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "sentence-transformers>=2.7.0",
    "transformers>=4.51.0",
    "torch>=2.2.0",
    "numpy>=1.26.0",
]
